Transformers
Created an instruct-based model with 111 million parameters, viewable here: https://huggingface.co/Thalesian/akk-111m. Code will automatically pull from this huggingface repo

As an instruct model, this allows for the same model to handle translation and transcription tasks. Bleu scores are not calculated exactly the same way as in the paper - any guidance on this would help. This platform is expandable to much larger instruct models as well.